pack and unpack

+++ module/vs2019/wpf/FileControl.xaml.cs
@@ -5,8 +5,10 @@ using System.IO;
 using System.Text.Encodings.Web;
 using System.Text.Json;
 using System.Windows.Controls;
+using Forms = System.Windows.Forms;
 using System.Windows.Media.Imaging;
 using XTC.oelArchive;
+using System.Collections.Generic;
 
 namespace oel.archive
 {
@@ -29,6 +31,7 @@ namespace oel.archive
         public FileFacade facade { get; set; }
 
         private FileReader reader_ { get; set; }
+        private FileWriter writer_ { get; set; }
         private string path_ { get; set; }
 
         public FileControl()
@@ -66,7 +69,10 @@ namespace oel.archive
                 reader_.Open(path_);
                 foreach (string entry in reader_.entries)
                 {
-                    lbEntry.Items.Add(entry);
+                    ListBoxItem item = new ListBoxItem();
+                    item.Content = entry;
+                    item.Uid = entry;
+                    lbEntry.Items.Add(item);
                 }
                 btnOpen.Visibility = System.Windows.Visibility.Collapsed;
                 btnCreate.Visibility = System.Windows.Visibility.Collapsed;
@@ -82,9 +88,11 @@ namespace oel.archive
 
         private void onEntrySelectionChanged(object sender, SelectionChangedEventArgs e)
         {
-            string entry = lbEntry.SelectedItem as string;
-            if (string.IsNullOrEmpty(entry))
+            var item = lbEntry.SelectedItem as ListBoxItem;
+            if (null == item)
                 return;
+
+            string entry = item.Content as string;
 
             spInfo.Visibility = System.Windows.Visibility.Visible;
             tbEntry.Text = entry;
@@ -95,31 +103,64 @@ namespace oel.archive
                 if (entry.EndsWith(".png"))
                 {
                     //不能适用using方式,否则图片的数据会被释放
-                    Stream ms = new MemoryStream(reader_.Read(entry));
-                    PngBitmapDecoder decoder = new PngBitmapDecoder(ms, BitmapCreateOptions.PreservePixelFormat, BitmapCacheOption.Default);
-                    var frame = decoder.Frames[0];
-                    imgViewer.ImageSource = frame;
-                    tbTip.Text = string.Format("{0}x{1}   {2}", frame.PixelWidth, frame.PixelHeight, formatSize(ms.Length));
+                    Stream ms = null;
+                    if (null != reader_)
+                    {
+                        ms = new MemoryStream(reader_.Read(entry));
+                    }
+                    else if (null != writer_)
+                    {
+                        ms = new MemoryStream(File.ReadAllBytes(Path.Combine(path_, entry)));
+                    }
+                    if (null != ms)
+                    {
+                        PngBitmapDecoder decoder = new PngBitmapDecoder(ms, BitmapCreateOptions.PreservePixelFormat, BitmapCacheOption.Default);
+                        var frame = decoder.Frames[0];
+                        imgViewer.ImageSource = frame;
+                        tbTip.Text = string.Format("{0}x{1}   {2}", frame.PixelWidth, frame.PixelHeight, formatSize(ms.Length));
+                    }
                 }
                 else if (entry.EndsWith(".jpg"))
                 {
                     //不能适用using方式,否则图片的数据会被释放
-                    Stream ms = new MemoryStream(reader_.Read(entry));
-                    JpegBitmapDecoder decoder = new JpegBitmapDecoder(ms, BitmapCreateOptions.PreservePixelFormat, BitmapCacheOption.Default);
-                    var frame = decoder.Frames[0];
-                    imgViewer.ImageSource = frame;
-                    tbTip.Text = string.Format("{0}x{1}   {2}", frame.PixelWidth, frame.PixelHeight, formatSize(ms.Length));
+                    Stream ms = null;
+                    if (null != reader_)
+                    {
+                        ms = new MemoryStream(reader_.Read(entry));
+                    }
+                    else if (null != writer_)
+                    {
+                        ms = new MemoryStream(File.ReadAllBytes(Path.Combine(path_, entry)));
+                    }
+                    if (null != ms)
+                    {
+                        JpegBitmapDecoder decoder = new JpegBitmapDecoder(ms, BitmapCreateOptions.PreservePixelFormat, BitmapCacheOption.Default);
+                        var frame = decoder.Frames[0];
+                        imgViewer.ImageSource = frame;
+                        tbTip.Text = string.Format("{0}x{1}   {2}", frame.PixelWidth, frame.PixelHeight, formatSize(ms.Length));
+                    }
                 }
                 else if (entry.EndsWith(".json"))
                 {
-                    using (var ms = new MemoryStream())
+                    JsonDocument doc = null;
+                    if (null != reader_)
                     {
-                        var document = JsonDocument.Parse(reader_.Read(entry));
-                        var writer = new Utf8JsonWriter(ms, new JsonWriterOptions { Indented = true, Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping });
-                        document.WriteTo(writer);
-                        writer.Flush();
-                        txtViewer.Text = System.Text.Encoding.UTF8.GetString(ms.ToArray());
-                        tbTip.Text = string.Format("{0}", formatSize(ms.Length));
+                        doc = JsonDocument.Parse(reader_.Read(entry));
+                    }
+                    else
+                    {
+                        doc = JsonDocument.Parse(File.ReadAllBytes(Path.Combine(path_, entry)));
+                    }
+                    if (null != doc)
+                    {
+                        using (var ms = new MemoryStream())
+                        {
+                            var writer = new Utf8JsonWriter(ms, new JsonWriterOptions { Indented = true, Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping });
+                            doc.WriteTo(writer);
+                            writer.Flush();
+                            txtViewer.Text = System.Text.Encoding.UTF8.GetString(ms.ToArray());
+                            tbTip.Text = string.Format("{0}", formatSize(ms.Length));
+                        }
                     }
                 }
             }
@@ -131,7 +172,16 @@ namespace oel.archive
 
         private void onCloseClicked(object sender, System.Windows.RoutedEventArgs e)
         {
-            reader_.Close();
+            if (null != reader_)
+            {
+                reader_.Close();
+                reader_ = null;
+            }
+            if (null != writer_)
+            {
+                writer_.Close();
+                writer_ = null;
+            }
             lbEntry.Items.Clear();
             spInfo.Visibility = System.Windows.Visibility.Hidden;
             btnOpen.Visibility = System.Windows.Visibility.Visible;
@@ -143,32 +193,68 @@ namespace oel.archive
 
         private void onUnpackClicked(object sender, System.Windows.RoutedEventArgs e)
         {
-            string dir = Path.GetDirectoryName(path_);
+            Forms.FolderBrowserDialog dialog = new Forms.FolderBrowserDialog();
+            if (Forms.DialogResult.OK != dialog.ShowDialog())
+                return;
+
+            string dir = dialog.SelectedPath;
             dir = Path.Combine(dir, Path.GetFileNameWithoutExtension(path_));
             if (Directory.Exists(dir))
                 Directory.Delete(dir, true);
             Directory.CreateDirectory(dir);
 
-            foreach(var entry in reader_.entries)
+            foreach (var entry in reader_.entries)
             {
-                string filepath = Path.Combine(dir, entry);
                 string subdir = Path.Combine(dir, Path.GetDirectoryName(entry));
+                if (!Directory.Exists(subdir))
+                    Directory.CreateDirectory(subdir);
+                string filepath = Path.Combine(dir, entry);
                 File.WriteAllBytes(filepath, reader_.Read(entry));
             }
         }
 
         private void onPackClicked(object sender, System.Windows.RoutedEventArgs e)
         {
-
+            SaveFileDialog dialog = new SaveFileDialog();
+            if (false == dialog.ShowDialog())
+                return;
+            string file = dialog.FileName;
+            if (!string.IsNullOrEmpty(tbPassword.Text))
+                writer_.SetPassword(tbPassword.Text);
+            writer_.Open(file, true);
+            foreach (var item in lbEntry.Items)
+            {
+                var lbItem = item as ListBoxItem;
+                writer_.Write((string)lbItem.Content, File.ReadAllBytes(lbItem.Uid));
+            }
+            writer_.Flush();
         }
 
         private void onCreateClicked(object sender, System.Windows.RoutedEventArgs e)
         {
+            Forms.FolderBrowserDialog dialog = new Forms.FolderBrowserDialog();
+            if (Forms.DialogResult.OK != dialog.ShowDialog())
+                return;
+
+            path_ = dialog.SelectedPath;
+
             btnOpen.Visibility = System.Windows.Visibility.Collapsed;
             btnCreate.Visibility = System.Windows.Visibility.Collapsed;
             btnPack.Visibility = System.Windows.Visibility.Visible;
             btnUnpack.Visibility = System.Windows.Visibility.Collapsed;
             btnClose.Visibility = System.Windows.Visibility.Visible;
+
+            List<string> files = new List<string>();
+            getAllFiles(path_, path_, ref files);
+            foreach (var file in files)
+            {
+                ListBoxItem item = new ListBoxItem();
+                item.Content = file;
+                item.Uid = Path.Combine(path_, file);
+                lbEntry.Items.Add(item);
+            }
+
+            writer_ = new FileWriter();
         }
 
         private string formatSize(long _size)
@@ -180,6 +266,23 @@ namespace oel.archive
             if (_size < 1024 * 1024 * 1024)
                 return string.Format("{0} MB", _size / 1024 / 1024);
             return string.Format("{0} GB", _size / 1024 / 1024 / 1024);
+        }
+
+        private void getAllFiles(string _dir, string _subdir, ref List<string> _files)
+        {
+            DirectoryInfo folder = new DirectoryInfo(_subdir);
+            DirectoryInfo[] di = folder.GetDirectories();
+            FileInfo[] fi = folder.GetFiles();
+
+            foreach (FileInfo file in fi)
+            {
+                string path = Path.GetRelativePath(_dir, file.FullName);
+                _files.Add(path);
+            }
+            foreach (DirectoryInfo subdir in di)
+            {
+                getAllFiles(_dir, subdir.FullName, ref _files);
+            }
         }
     }
 }
